chore: move gapic option func to gapic.go
