+++ internal/librarian/python/generate_test.go
@@ -1428,111 +1428,6 @@ func TestCreateRepoMetadata_Error(t *testing.T) {
 	}
 }
 
-func TestDeriveGAPICNamespace(t *testing.T) {
-	for _, test := range []struct {
-		name string
-		path string
-		want string
-	}{
-		{
-			name: "single path element",
-			path: "grafeas",
-			want: "grafeas",
-		},
-		{
-			name: "single path element with version",
-			path: "grafeas/v1",
-			want: "grafeas",
-		},
-		{
-			name: "multiple path elements",
-			path: "google/cloud/datacatalog/lineage/v1",
-			want: "google.cloud",
-		},
-	} {
-		t.Run(test.name, func(t *testing.T) {
-			got := deriveGAPICNamespace(test.path)
-			if diff := cmp.Diff(test.want, got); diff != "" {
-				t.Errorf("mismatch (-want +got):\n%s", diff)
-			}
-		})
-	}
-}
-
-func TestDeriveGAPICName(t *testing.T) {
-	for _, test := range []struct {
-		name string
-		path string
-		want string
-	}{
-		{
-			name: "single path element in name",
-			path: "google/cloud/datacatalog/v1",
-			want: "datacatalog",
-		},
-		{
-			name: "multiple path elements in name",
-			path: "google/cloud/datacatalog/lineage/v1",
-			want: "datacatalog_lineage",
-		},
-		{
-			name: "no version",
-			path: "google/apps/script/type",
-			want: "script_type",
-		},
-	} {
-		t.Run(test.name, func(t *testing.T) {
-			got := deriveGAPICName(test.path)
-			if diff := cmp.Diff(test.want, got); diff != "" {
-				t.Errorf("mismatch (-want +got):\n%s", diff)
-			}
-		})
-	}
-}
-
-func TestFindOption(t *testing.T) {
-	for _, test := range []struct {
-		name      string
-		options   []string
-		wantValue string
-		wantOk    bool
-	}{
-		{
-			name:    "empty options",
-			options: []string{},
-		},
-		{
-			name:    "requested option not present",
-			options: []string{"a=b"},
-		},
-		{
-			name:    "requested option not present, but similar names are",
-			options: []string{"othertest=a", "testother=b"},
-		},
-		{
-			name:      "option present with value",
-			options:   []string{"a=b", "test=test-value", "c=d"},
-			wantValue: "test-value",
-			wantOk:    true,
-		},
-		{
-			name:    "option present without value",
-			options: []string{"a=b", "test=", "c=d"},
-			wantOk:  true,
-		},
-	} {
-		t.Run(test.name, func(t *testing.T) {
-			gotValue, gotOk := findOption(test.options, "test")
-			if diff := cmp.Diff(test.wantValue, gotValue); diff != "" {
-				t.Errorf("mismatch in value (-want +got):\n%s", diff)
-			}
-			if test.wantOk != gotOk {
-				t.Errorf("mismatch in found: want %v, got %v", test.wantOk, gotOk)
-			}
-		})
-	}
-}
-
 func TestCreateChangelog(t *testing.T) {
 	libName := "google-cloud-test"
 	output := t.TempDir()
